dio-pytest-agent/requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-openai>=0.1.7
openai>=1.48.0
pytest>=8.2.0
typer>=0.12.5
